First version of the PCB for a panel.
As tested with Peter on juli 2nd, 2011.

Note that during the making of the schematics, some (lots of) wires
were moved so the code in lasercutterpanel.pde had to be adapted
Also, the option to drive two leds has been added to the arduino code.
However, the new wire configuration still needs to be tested before 
ordering 200+ PCB's based on this!